requirements.txt
# Core Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Environment Configuration
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Data Science & ML
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2

# Deep Learning (TensorFlow)
tensorflow==2.15.0

# Visualization & Dashboard
streamlit==1.29.0
plotly==5.18.0

# Scheduling
APScheduler==3.10.4

# OpenAI Integration
openai==1.6.1

# SMS Notifications
twilio==8.11.0
boto3==1.34.10  # For AWS SNS

# NLP for PII Detection
spacy==3.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
flake8==6.1.0
black==23.12.1

# CORS Support
Flask-CORS==4.0.0

# Rate Limiting
Flask-Limiter==3.5.0

# Security
cryptography==41.0.7
